+++ protoc_plugin/test/high_tagnumber_test.dart
@@ -3,7 +3,9 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import '../out/protos/high_tagnumber.pb.dart';
+import '../out/protos/google/protobuf/empty.pb.dart';
 import 'package:test/test.dart';
+import 'package:fixnum/fixnum.dart';
 
 main() {
   test('round trip 29 bit tag number, binary encoding', () {
@@ -13,5 +15,12 @@ main() {
   test('round trip 29 bit tag number, jspblite2', () {
     expect(M.fromJson((M()..a = 43).writeToJson()), M()..a = 43);
     expect(M.fromJson((M()..b = 43).writeToJson()), M()..b = 43);
+  });
+  test('unknown fields', () {
+    final empty = Empty.fromBuffer((M()..a = 44).writeToBuffer());
+    expect(empty.unknownFields.isEmpty, false);
+    expect(empty.unknownFields.getField(M().info_.tagNumber('a')).varints,
+        [Int64(44)]);
+    expect(M.fromBuffer(empty.writeToBuffer()).a, 44);
   });
 }